+++ src/realm/utilities.cpp
@@ -258,7 +258,7 @@ uint64_t fastrand(uint64_t max, bool is_seed)
 }
 
 
-void millisleep(size_t milliseconds)
+void millisleep(unsigned long milliseconds)
 {
 #ifdef _WIN32
     _sleep(milliseconds);
