+++ test/test_query.cpp
@@ -3021,8 +3021,8 @@ TEST(Query_DoubleCoordinates)
 
     for (size_t t = 0; t < 100000; t++) {
         table->add_empty_row(1);
-        table->set_double(0, t, (t * 12345) % 1000);
-        table->set_double(1, t, (t * 12345) % 1000);
+        table->set_double(0, t, double((t * 12345) % 1000));
+        table->set_double(1, t, double((t * 12345) % 1000));
 
         if (table->get_double(0, t) >= 100. && table->get_double(0, t) <= 110. && table->get_double(1, t) >= 100. &&
             table->get_double(1, t) <= 110.) {
